report-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
google-cloud-firestore==2.13.1
python-dateutil==2.8.2
